Some more changes by Ajay
