+++ Frontend/src/Components/Admin/Landing Comp/ContactIcon.css
@@ -1,7 +1,3 @@
-body {
-  padding-bottom: 150px; /* Increase this value if needed */
-}
-
 .contact-icon-container {
   z-index: 1001;
   position: fixed;
